Refactor 10: optimizer
